ref: improve code readability
