+++ src/multiplayer/multiplayer.js
@@ -4,8 +4,8 @@ import { getNodeIdGenerator } from '../NodeIdGenerator';
 import { getDefaultAnchorPoint } from '@/portsUtils';
 import Room from './room';
 import store from '@/store';
-import { setEventTimerDefinition } from '@/components/nodes/boundaryTimerEvent';
 import { getBoundaryEventData } from '@/components/nodes/boundaryEvent/boundaryEventUtils';
+import { InspectorUtils } from './inspector.utils';
 
 export default class Multiplayer {
   clientIO = null;
@@ -14,6 +14,7 @@ export default class Multiplayer {
   modeler = null;
   #nodeIdGenerator = null;
   room = null;
+  inspector = null;
   deletedItem = null;
 
   constructor(modeler) {
@@ -30,6 +31,7 @@ export default class Multiplayer {
     // Get the room name from the process id
     const processId = window.ProcessMaker.modeler.process.uuid ?? window.ProcessMaker.modeler.process.id;
     this.room = new Room(`room-${processId}`);
+    this.inspector = new InspectorUtils(this.modeler, store);
 
     // Connect to websocket server
     this.clientIO = io(window.ProcessMaker.multiplayer.host, { transports: ['websocket', 'polling']});
@@ -534,7 +536,7 @@ export default class Multiplayer {
   updateShapeFromInspector(data) {
     let node = null;
     if (data.oldNodeId && data.oldNodeId !== data.id) {
-      this.updateNodeId(data.oldNodeId, data.id);
+      this.inspector.updateNodeId(data.oldNodeId, data.id);
     }
 
     node = this.getNodeById(data.id);
@@ -546,7 +548,7 @@ export default class Multiplayer {
     const { definition } = node;
 
     if (data.loopCharacteristics) {
-      this.handleLoopCharacteristics(node, data.loopCharacteristics);
+      this.inspector.handleLoopCharacteristics(node, data.loopCharacteristics);
       return;
     }
 
@@ -560,111 +562,22 @@ export default class Multiplayer {
     const value = data[key];
 
     if (key === 'condition') {
-      this.updateEventCondition(definition, value);
+      this.inspector.updateEventCondition(definition, value);
     } else if (key === 'gatewayDirection') {
-      this.updateGatewayDirection(definition, value);
+      this.inspector.updateGatewayDirection(definition, value);
     } else if (key === 'messageRef') {
-      this.updateMessageRef(node, value, extras);
+      this.inspector.updateMessageRef(node, value, extras);
     } else if (key === 'signalRef') {
-      this.updateSignalRef(node, value, extras);
+      this.inspector.updateSignalRef(node, value, extras);
     } else if (key === 'signalPayload') {
-      this.updateSignalPayload(node, value);
+      this.inspector.updateSignalPayload(node, value);
     } else if (key === 'eventTimerDefinition') {
-      this.updateEventTimerDefinition(node, value);
-    } else if (!this.isSpecialProperty(key)) {
-      this.updateNodeProperty(node, key, value);
+      this.inspector.updateEventTimerDefinition(node, value);
+    } else if (!this.inspector.isSpecialProperty(key)) {
+      this.inspector.updateNodeProperty(node, key, value);
     }
   }
 
-  updateNodeId(oldNodeId, newId) {
-    const index = this.getIndex(oldNodeId);
-    const yNode = this.yArray.get(index);
-    yNode.set('id', newId);
-    const node = this.getNodeById(oldNodeId);
-    store.commit('updateNodeProp', { node, key: 'id', value: newId });
-  }
-
-  handleLoopCharacteristics(node, loopCharacteristics) {
-    const parsedLoopCharacteristics = JSON.parse(loopCharacteristics);
-    this.modeler.nodeRegistry[node.type].loopCharacteristicsHandler({
-      type: node.definition.type,
-      '$loopCharactetistics': {
-        id: node.id,
-        loopCharacteristics: parsedLoopCharacteristics,
-      },
-    }, node, this.setNodeProp, this.modeler.moddle, this.modeler.definitions, false);
-  }
-
-  updateEventCondition(definition, value) {
-    definition.get('eventDefinitions')[0].get('condition').body = value;
-  }
-
-  updateGatewayDirection(definition, value) {
-    definition.set('gatewayDirection', value);
-  }
-
-  isSpecialProperty(key) {
-    const specialProperties = [
-      'messageRef', 'signalRef', 'signalPayload', 'gatewayDirection', 'condition', 'allowedUsers', 'allowedGroups',
-    ];
-    return specialProperties.includes(key);
-  }
-
-  updateNodeProperty(node, key, value) {
-    store.commit('updateNodeProp', { node, key, value });
-  }
-  updateMessageRef(node, value, extras) {
-    let message = this.modeler.definitions.rootElements.find(element => element.id === value);
-
-    if (!message) {
-      message = this.modeler.moddle.create('bpmn:Message', {
-        id: value,
-        name: extras?.messageName || value,
-      });
-      this.modeler.definitions.rootElements.push(message);
-    }
-
-    node.definition.get('eventDefinitions')[0].messageRef = message;
-
-    if (extras?.allowedUsers) {
-      node.definition.set('allowedUsers', extras.allowedUsers);
-    }
-
-    if (extras?.allowedGroups) {
-      node.definition.set('allowedGroups', extras.allowedGroups);
-    }
-  }
-
-  updateSignalRef(node, value, extras) {
-    let signal = this.modeler.definitions.rootElements.find(element => element.id === value);
-
-    if (!signal) {
-      signal = this.modeler.moddle.create('bpmn:Signal', {
-        id: value,
-        name: extras?.signalName || value,
-      });
-      this.modeler.definitions.rootElements.push(signal);
-    }
-
-    node.definition.get('eventDefinitions')[0].signalRef = signal;
-  }
-
-  updateSignalPayload(node, value) {
-    const eventDefinitions = node.definition.get('eventDefinitions');
-    const SIGNAL_EVENT_DEFINITION_TYPE = 'bpmn:SignalEventDefinition';
-
-    eventDefinitions.forEach(definition => {
-      if (definition.$type === SIGNAL_EVENT_DEFINITION_TYPE) {
-        definition.config = value;
-      }
-    });
-  }
-
-  updateEventTimerDefinition(node, value) {
-    const { type, body } = value;
-    const eventDefinitions = setEventTimerDefinition(this.modeler.moddle, node, type, body);
-    store.commit('updateNodeProp', { node, key: 'eventDefinitions', value: eventDefinitions });
-  }
   /**
    * Update the shared document and emit socket sign to update the flows
    * @param {Object} data
